13-clients/cli/pye-console/requirements.txt
click>=8.1.0
requests>=2.31.0
websocket-client>=1.6.0
rich>=13.0.0
pydantic>=2.0.0
cryptography>=41.0.0
